More small cleanup

+++ src/overlays/gamestates/ovl_select/z_select.c
@@ -45,7 +45,7 @@ void MapSelect_LoadGame(MapSelectState* this, u32 entranceIndex, s32 spawnIndex)
     }
 
     gSaveContext.respawn[RESPAWN_MODE_DOWN].entranceIndex = 0xFFFF;
-    gSaveContext.seqIndex = (u8)NA_BGM_DISABLED;
+    gSaveContext.seqIndex = NA_BGM_DISABLED;
     gSaveContext.nightSeqIndex = 0xFF;
     gSaveContext.showTitleCard = true;
     gSaveContext.respawnFlag = 0;
